add left and right padding
To give left and right space on smaller devices, I added padding in the
CSS. First, I added left and right padding to the logo, navigation and
post entries. Next, I added left padding to the post titles, dates and
sidebar. Finally, I added right padding to the post author.

I was going to try to use media queries to do this, adding it to the CSS
was more efficient than adding several media queries or using max-width,
with is generally a no-no.